requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
jinja2==3.1.4
aiosmtplib==3.0.2
apscheduler==3.10.4
email-validator==2.2.0
pyyaml==6.0.2
